pyproject.toml
[project]
name = "ray-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "openai>=1.107.3",
    "pillow>=11.3.0",
    "ray[default,serve,server]==2.48.0",
    "requests>=2.32.5",
    "torch>=1.6.0",
    "transformers>=4.0.0",
]
